// src/data/contracts/apis/facebook.ts
export interface LoadFacebookUserApi {
  loadUser: (
    params: LoadFacebookUserApi.Params
  ) => Promise<LoadFacebookUserApi.Result>;
}

export namespace LoadFacebookUserApi {
  export type Params = {
    token: string;
  };

  export type Result =
    | {
        facebookId: string;
        email: string;
        name: string;
      }
    | undefined;
}